Rename WrappedClient into AsyncHTTPClient
